requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
SQLAlchemy==2.0.49
psycopg2-binary==2.9.11
python-multipart==0.0.24
pydantic==2.12.5
starlette==1.0.0
python-dotenv